% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc2gains.R
\name{nc2gains}
\alias{nc2gains}
\title{nc2gains}
\usage{
nc2gains(file)
}
\arguments{
\item{file}{file to convert}
}
\value{
writes csv and returns a dataframe
}
\description{
writes csv files based on MAgPIE nc outputs
}
\author{
David Chen
}
